controller/requirements.txt
python-telegram-bot[webhooks]==21.*
redis[hiredis]==5.*
kubernetes==31.*
slack-bolt==1.*
